// lib/countries/reference.generated.ts
export interface CountryRef {
  m49: number;
  name: string;
  iso2: string;
  iso3: string;
  dial: string;
  currencyCode: string;
  currencyName: string;
  subRegionCode: number | null;
  subRegionName: string;
  regionName: string;
}
// ...
export const COUNTRY_REFERENCE_BY_M49: Map<number, CountryRef> = new Map(
  COUNTRY_REFERENCE.map((r) => [r.m49, r]),
);